ilabor360-scraper/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
undetected-chromedriver==3.5.5
flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
requests==2.31.0
